requirements.txt
absl-py==0.11.0
albumentations==0.5.2
appdirs==1.4.4
audioread==2.1.9
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
cycler==0.10.0
dataclasses==0.6
decorator==4.4.2
edafa==0.1.5
facenet-pytorch==2.5.1
future==0.18.2
google-auth==1.23.0
google-auth-oauthlib==0.4.2
grpcio==1.34.0
idna==2.10
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==3.1.1
joblib==0.17.0
kaldiio==2.17.0
kiwisolver==1.3.1
librosa==0.8.0
Markdown==3.3.3
matplotlib==3.3.3
networkx==2.5
numpy==1.19.4
oauthlib==3.1.0
opencv-python==4.5.1.48
opencv-python-headless==4.5.1.48
pandas==1.2.2
pathlib==1.0.1
Pillow==8.0.1
pooch==1.3.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.25.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.4.1
Shapely==1.7.1
six==1.15.0
sklearn==0.0
SoundFile==0.10.3.post1
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
threadpoolctl==2.1.0
tifffile==2020.12.4
torch==1.7.1
torchaudio
torchvision
tqdm==4.54.0
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
zipp==3.4.0
